made interactive_callbacks generic
